fix: Refractor and change  event carousel based on review
